Fix broken pipe error when connections are closed early by the client.
When configuring the maintenanceserver with haproxy, haproxy makes perodically an OPTIONS request for checking if the server still response.
This OPTIONS request is closed as soon as the first bytes arrive.
This results in exceptions such as:
```
  File "/usr/local/python/2.7.5-i386/lib/python2.7/socket.py", line 303, in flush
    self._sock.sendall(view[write_offset:write_offset+buffer_size])
error: [Errno 32] Broken pipe
```
This happens when we try to write to a closed connection.

The default request handler always tries to cleanup (`finish()`) after each request, even when exceptions happen during the request. `finish()` tries to write the remaining data to the stream and close it.

This behavior is usually correct, but when there is a broken pipe calling `finish()` will result in another socket error (broken pip) since it tries to write to the closed socket again.

This changes the behavior so that the request is not finished when we have a socket error (broken pipe).
This eliminates the disturbing exceptions from the log.