Avoid caching maintenance server responses.
Respond to GET, POST, and HEAD requests with 503 Service Unavailable in
order to avoid having the response cached in varnish or other cache
proxies.
OPTIONS is still responded with 200 OK, but with Cache-Control=no-cache,
because sending a 5xx-response for a OPTIONS-request will make HaProxy
to shut down the backend because it might be broken.